requirements.txt
flask 
werkzeug
python-docx
groq
deep-translator
aspose-pdf
pdf2docx
googletrans

